chore: Move common workers into a package
Starts adding a new package for common workers and refactors the
examples folder a little to start using multiple executors.
