+++ tierkreis_workers/tierkreis_workers/__init__.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+
+COMMON_WORKERS_REGISTRY = Path(__file__).parent
